requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
pytesseract==0.3.10
Pillow==10.0.0